Swapped webs don't trigger Defiant
Track the side that originally set Sticky Web, and if it matches the side the target of the speed drop is on, don't activate Defiant. Hard to tell if this is a bug or a feature in gen 8 - maybe the swapped webs are considered a "friendly" effect?

Also, fix Sticky Web not activating Defiant.